+++ config/jest.config.js
@@ -13,7 +13,7 @@ module.exports = (disabledMocks = []) => ({
     moduleNameMapper: {
         '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2)$': `${mockDir}/fileMock.ts`,
         '\\.(css|scss)$': `${mockDir}/emptyMock.ts`,
-        'pc-nrfjprog-js|nrf-device-setup|usb|applicationinsights': `${mockDir}/emptyMock.ts`,
+        applicationinsights: `${mockDir}/emptyMock.ts`,
         '^electron$': `${mockDir}/electronMock.ts`,
         '^electron-store$': `${mockDir}/electronStoreMock.ts`,
         '@electron/remote': `${mockDir}/remoteMock.ts`,